@@ -1,6 +1,6 @@
 {
   "name": "raf-website",
-  "version": "0.0.0",
+  "version": "0.1.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --open",
@@ -11,20 +11,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.9",
-    "@angular/common": "~9.1.9",
-    "@angular/compiler": "~9.1.9",
-    "@angular/core": "~9.1.9",
+    "@angular/animations": "~10.0.14",
+    "@angular/common": "~10.0.14",
+    "@angular/compiler": "~10.0.14",
+    "@angular/core": "~10.0.14",
     "@angular/fire": "^6.0.0",
-    "@angular/forms": "~9.1.9",
-    "@angular/platform-browser": "~9.1.9",
-    "@angular/platform-browser-dynamic": "~9.1.9",
-    "@angular/router": "~9.1.9",
+    "@angular/forms": "~10.0.14",
+    "@angular/platform-browser": "~10.0.14",
+    "@angular/platform-browser-dynamic": "~10.0.14",
+    "@angular/router": "~10.0.14",
     "@fortawesome/angular-fontawesome": "^0.6.1",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-brands-svg-icons": "^5.13.0",
     "@fortawesome/free-regular-svg-icons": "^5.13.0",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@sendgrid/mail": "^7.2.5",
+    "bulma": "^0.9.1",
+    "cors": "^2.8.5",
     "firebase": "^7.14.6",
     "materialize-css": "^1.0.0",
     "rxjs": "~6.5.4",
@@ -32,9 +35,9 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.7",
-    "@angular/cli": "~9.1.7",
-    "@angular/compiler-cli": "~9.1.9",
+    "@angular-devkit/build-angular": "~0.1000.8",
+    "@angular/cli": "~10.0.8",
+    "@angular/compiler-cli": "~10.0.14",
     "@types/node": "^12.11.1",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
@@ -49,7 +52,7 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.8.3",
+    "typescript": "~3.9.7",
     "@angular-devkit/architect": "~0.900",
     "firebase-tools": "^8.0.0",
     "fuzzy": "^0.1.3",
